requirements.txt
python=3.9
numpy
scipy
pandas
scikit-learn
pytorch
spacy
spacy-fastlang
spacy-language-detection
transformers
flask
flask_cors
